updated header file

+++ 0x14-bit_manipulation/holberton.h
@@ -11,5 +11,6 @@ int set_bit(unsigned long int *n, unsigned int index);
 int clear_bit(unsigned long int *n, unsigned int index);
 int _putchar(char c);
 unsigned int flip_bits(unsigned long int n, unsigned long int m);
-int countSetBits(int n);
+/*int countSetBits(int n);*/
+unsigned int countSetBits(unsigned long int n);
 #endif
